move querydb classes into querydb namespace: q, Query, QueryDatabase
