+++ querydb/src/main/scala/io/joern/dumpq/Main.scala
@@ -1,6 +1,6 @@
 package io.joern.dumpq
 
-import io.joern.console.{DefaultArgumentProvider, QueryDatabase}
+import io.joern.querydb.{DefaultArgumentProvider, QueryDatabase}
 import io.joern.dataflowengineoss.queryengine.{EngineConfig, EngineContext}
 import io.joern.dataflowengineoss.semanticsloader.Semantics
 import org.json4s.{Formats, NoTypeHints}
