+++ querydb/src/main/scala/io/joern/scanners/c/CopyLoops.scala
@@ -2,6 +2,7 @@ package io.joern.scanners.c
 
 import io.joern.scanners._
 import io.joern.console._
+import io.joern.querydb._
 import io.joern.macros.QueryMacros._
 import io.shiftleft.semanticcpg.language._
