+++ querydb/src/main/scala/io/joern/scanners/c/FileOpRace.scala
@@ -3,6 +3,7 @@ package io.joern.scanners.c
 import io.joern.scanners.{Crew, QueryTags}
 import io.shiftleft.codepropertygraph.generated.nodes._
 import io.joern.console._
+import io.joern.querydb._
 import io.joern.dataflowengineoss.queryengine.EngineContext
 import io.shiftleft.semanticcpg.language._
 import io.joern.macros.QueryMacros._
