+++ querydb/src/main/scala/io/joern/scanners/c/MissingLengthCheck.scala
@@ -3,6 +3,7 @@ package io.joern.scanners.c
 import io.shiftleft.semanticcpg.language.{ICallResolver, NoResolve}
 import io.joern.scanners._
 import io.joern.console._
+import io.joern.querydb._
 import io.shiftleft.codepropertygraph.generated.nodes
 import io.joern.dataflowengineoss.queryengine.EngineContext
 import io.shiftleft.semanticcpg.language._
