+++ querydb/src/main/scala/io/joern/scanners/c/NullTermination.scala
@@ -4,6 +4,7 @@ import io.joern.scanners.{Crew, QueryTags}
 import io.shiftleft.semanticcpg.language._
 import io.joern.dataflowengineoss.language._
 import io.joern.console._
+import io.joern.querydb._
 import io.joern.dataflowengineoss.queryengine.EngineContext
 import io.joern.dataflowengineoss.semanticsloader.Semantics
 import io.joern.macros.QueryMacros._
