+++ querydb/src/main/scala/io/joern/scanners/c/SignedLeftShift.scala
@@ -3,6 +3,7 @@ package io.joern.scanners.c
 import io.joern.scanners._
 import io.shiftleft.codepropertygraph.generated.Operators
 import io.joern.console._
+import io.joern.querydb._
 import io.joern.macros.QueryMacros._
 import io.shiftleft.semanticcpg.language._
