+++ querydb/src/main/scala/io/joern/scanners/ghidra/UserInputIntoDangerousFunctions.scala
@@ -2,6 +2,7 @@ package io.joern.scanners.ghidra
 
 import io.joern.scanners._
 import io.joern.console._
+import io.joern.querydb._
 import io.joern.macros.QueryMacros._
 import io.shiftleft.semanticcpg.language._
 import io.joern.dataflowengineoss.language._
