+++ querydb/src/main/scala/io/joern/scanners/java/DangerousFunctions.scala
@@ -3,6 +3,7 @@ package io.joern.scanners.java
 import io.joern.scanners._
 import io.shiftleft.semanticcpg.language._
 import io.joern.console._
+import io.joern.querydb._
 import io.joern.macros.QueryMacros._
 
 object DangerousFunctions extends QueryBundle {
