+++ querydb/src/test/scala/io/joern/scanners/c/QueryWithReachableBy.scala
@@ -1,7 +1,7 @@
 package io.joern.scanners.c
 
 import io.joern.scanners._
-import io.joern.console._
+import io.joern.querydb._
 import io.joern.dataflowengineoss.language._
 import io.shiftleft.semanticcpg.language._
 import io.joern.macros.QueryMacros._
