+++ querydb/src/test/scala/io/joern/suites/AllBundlesTestSuite.scala
@@ -1,6 +1,6 @@
 package io.joern.suites
 
-import io.joern.console.QueryDatabase
+import io.joern.querydb.QueryDatabase
 import org.scalatest.wordspec.AnyWordSpec
 import org.scalatest.matchers.should.Matchers.convertToAnyShouldWrapper
