+++ querydb/src/test/scala/io/joern/suites/CQueryTestSuite.scala
@@ -3,8 +3,8 @@ package io.joern.suites
 import io.joern.util.QueryUtil
 import io.shiftleft.codepropertygraph.generated.nodes
 import io.joern.console.scan._
-import io.joern.console.QueryBundle
-import io.joern.console.Query
+import io.joern.querydb.QueryBundle
+import io.joern.querydb.Query
 import io.joern.c2cpg.testfixtures.DataFlowCodeToCpgSuite
 import io.joern.x2cpg.testfixtures.TestCpg
 import io.shiftleft.semanticcpg.language._
