+++ querydb/src/test/scala/io/joern/suites/GhidraQueryTestSuite.scala
@@ -1,11 +1,11 @@
 package io.joern.suites
 
-import io.joern.console.QueryBundle
+import io.joern.querydb.QueryBundle
 import io.joern.console.scan._
 import io.joern.ghidra2cpg.fixtures.DataFlowBinToCpgSuite
 import io.joern.util.QueryUtil
 import io.shiftleft.codepropertygraph.generated.nodes
-import io.joern.console.Query
+import io.joern.querydb.Query
 import io.shiftleft.semanticcpg.language._
 import io.shiftleft.utils.ProjectRoot
 import overflowdb.traversal.iterableToTraversal
