+++ querydb/src/test/scala/io/joern/suites/JavaQueryTestSuite.scala
@@ -1,7 +1,7 @@
 package io.joern.suites
 
 import io.joern.console.scan._
-import io.joern.console.{CodeSnippet, Query, QueryBundle}
+import io.joern.querydb.{CodeSnippet, Query, QueryBundle}
 import io.joern.javasrc2cpg.testfixtures.JavaSrcCode2CpgFixture
 import io.joern.util.QueryUtil
 import io.joern.x2cpg.testfixtures.TestCpg
