+++ querydb/src/test/scala/io/joern/util/QueryUtil.scala
@@ -1,8 +1,8 @@
 package io.joern.util
 
 import io.joern.suites.QDBArgumentProvider
-import io.joern.console.{QueryBundle, QueryDatabase}
-import io.joern.console.Query
+import io.joern.querydb.{QueryBundle, QueryDatabase}
+import io.joern.querydb.Query
 
 object QueryUtil {
   object EmptyBundle extends QueryBundle
